+++ cosmos/models/mobility/src/main/java/es/tid/cosmos/mobility/mapreduce/ParseCdrsMapper.java
@@ -3,7 +3,6 @@ package es.tid.cosmos.mobility.mapreduce;
 import java.io.IOException;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -15,7 +14,7 @@ import es.tid.cosmos.mobility.data.MobProtocol.Cdr;
  *
  * @author sortega
  */
-public class ParseCdrsMapper extends Mapper<IntWritable, Text, LongWritable,
+public class ParseCdrsMapper extends Mapper<LongWritable, Text, LongWritable,
         ProtobufWritable<Cdr>> {
     private LongWritable userId;
     
@@ -25,7 +24,7 @@ public class ParseCdrsMapper extends Mapper<IntWritable, Text, LongWritable,
     }
     
     @Override
-    public void map(IntWritable key, Text line, Context context)
+    public void map(LongWritable key, Text line, Context context)
             throws IOException, InterruptedException {
         ProtobufWritable<Cdr> cdr = CdrUtil.wrap(CdrUtil.parse(line.toString()));
         this.userId.set(cdr.get().getUserId());
